refactor(api): extract summarize-progress to its own module
progress.ts exceeded the 400-line limit after the correction-swap
guard was added. Move updateSummarizeProgress and its two wrappers
to summarize-progress.ts and re-export from progress.ts to keep
existing imports unbroken.
